refactor(core): Move validate package to internal
As we deferred any anything except our proof-of-concept validation,
this is definitely something we should keep as an implementation
detail until we've thought more about it. (We potentially should
even delete it as it's a bit strange that we're only validating
one field, but at least let's make it internal.)

Also fix the lint failure, which doesn't like repetition in
package/function names (validate.ValidateHalConfig).

Finally change some names to lowercase so they aren't exported
as they aren't used outside the package.
